Theme Changes & Some Functions (see desc)
Created a navbar folder under components.
Created navbar.css to take care of navbar specific styles.
Changed navbar background to a darker color, added hover options.

Added icons in actions

Noticed that I forgot to change amount from integer to float
(Changed it manually in database but added it migrations)
Added get value function to make it look better if a number has decimals

+++ database/migrations/2024_03_01_170750_create_expense_income_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('expense_income', function (Blueprint $table) {
             $table->id();
-            $table->integer("amount");
+            $table->float("amount");
             $table->date("time");
             $table->string("description")->nullable();
             $table->unsignedBigInteger("type_id");
